Singular component for rendering page list
