+++ pages/api/session/login.ts
@@ -1,6 +1,5 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import nc from 'next-connect';
-import { User } from '@prisma/client';
 import { prisma } from 'db';
 import { onError, onNoMatch } from 'lib/middleware';
 import { withSessionRoute } from 'lib/session/withSession';
